Reorganise files in bias metrics folder
