+++ frontend/src/pages/modelServing/screens/metrics/bias/BiasConfigurationPage/tableData.tsx
